# Improve efficiency of the `inv_link_<ordinal_family>()` functions