--- a/R/distributions.R
+++ b/R/distributions.R
@@ -2096,8 +2096,10 @@ inv_link_sratio <- function(x, link) {
   dim_noncat <- dim(x)[-ndim]
   marg_othdim <- seq_along(dim(x))[-ndim]
   ones_arr <- array(1, dim = c(dim_noncat, 1))
-  Sx_cumprod <- aperm(apply(1 - x, marg_othdim, cumprod),
-                      perm = c(marg_othdim + 1, 1))
+  Sx_cumprod <- aperm(
+    apply(1 - x, marg_othdim, cumprod),
+    perm = c(marg_othdim + 1, 1)
+  )
   abind::abind(x, ones_arr) * abind::abind(ones_arr, Sx_cumprod)
 }
 
@@ -2146,8 +2148,10 @@ inv_link_cratio <- function(x, link) {
   dim_noncat <- dim(x)[-ndim]
   marg_othdim <- seq_along(dim(x))[-ndim]
   ones_arr <- array(1, dim = c(dim_noncat, 1))
-  x_cumprod <- aperm(apply(x, marg_othdim, cumprod),
-                     perm = c(marg_othdim + 1, 1))
+  x_cumprod <- aperm(
+    apply(x, marg_othdim, cumprod),
+    perm = c(marg_othdim + 1, 1)
+  )
   abind::abind(1 - x, ones_arr) * abind::abind(ones_arr, x_cumprod)
 }
 
@@ -2193,18 +2197,25 @@ inv_link_acat <- function(x, link) {
   ones_arr <- array(1, dim = c(dim_noncat, 1))
   if (link == "logit") { 
     # faster evaluation in this case
-    exp_x_cumprod <- aperm(apply(exp(x), marg_othdim, cumprod),
-                           perm = c(marg_othdim + 1, 1))
+    exp_x_cumprod <- aperm(
+      apply(exp(x), marg_othdim, cumprod),
+      perm = c(marg_othdim + 1, 1)
+    )
     out <- abind::abind(ones_arr, exp_x_cumprod)
   } else {
     x <- ilink(x, link)
-    x_cumprod <- aperm(apply(x, marg_othdim, cumprod),
-                       perm = c(marg_othdim + 1, 1))
+    x_cumprod <- aperm(
+      apply(x, marg_othdim, cumprod),
+      perm = c(marg_othdim + 1, 1)
+    )
     nthres <- dim(x)[ndim]
-    Sx_cumprod_rev <- aperm(apply(
-      1 - slice(x, ndim, rev(seq_len(nthres)), drop = FALSE),
-      marg_othdim, cumprod
-    ), perm = c(marg_othdim + 1, 1))
+    Sx_cumprod_rev <- aperm(
+      apply(
+        1 - slice(x, ndim, rev(seq_len(nthres)), drop = FALSE),
+        marg_othdim, cumprod
+      ), 
+      perm = c(marg_othdim + 1, 1)
+    )
     Sx_cumprod_rev <- slice(
       Sx_cumprod_rev, ndim, rev(seq_len(nthres)), drop = FALSE
     )