--- a/R/distributions.R
+++ b/R/distributions.R
@@ -2097,8 +2097,9 @@ inv_link_sratio <- function(x, link) {
   dim_thres <- dim(x)[ndim]
   marg_othdim <- seq_along(dim(x))[-ndim]
   ones_arr <- array(1, dim = c(dim_noncat, 1))
+  dim_t <- c(dim_thres, dim_noncat)
   Sx_cumprod <- aperm(
-    array(apply(1 - x, marg_othdim, cumprod), dim = c(dim_thres, dim_noncat)),
+    array(apply(1 - x, marg_othdim, cumprod), dim = dim_t),
     perm = c(marg_othdim + 1, 1)
   )
   abind::abind(x, ones_arr) * abind::abind(ones_arr, Sx_cumprod)
@@ -2150,8 +2151,9 @@ inv_link_cratio <- function(x, link) {
   dim_thres <- dim(x)[ndim]
   marg_othdim <- seq_along(dim(x))[-ndim]
   ones_arr <- array(1, dim = c(dim_noncat, 1))
+  dim_t <- c(dim_thres, dim_noncat)
   x_cumprod <- aperm(
-    array(apply(x, marg_othdim, cumprod), dim = c(dim_thres, dim_noncat)),
+    array(apply(x, marg_othdim, cumprod), dim = dim_t),
     perm = c(marg_othdim + 1, 1)
   )
   abind::abind(1 - x, ones_arr) * abind::abind(ones_arr, x_cumprod)
@@ -2198,25 +2200,27 @@ inv_link_acat <- function(x, link) {
   dim_thres <- dim(x)[ndim]
   marg_othdim <- seq_along(dim(x))[-ndim]
   ones_arr <- array(1, dim = c(dim_noncat, 1))
+  dim_t <- c(dim_thres, dim_noncat)
   if (link == "logit") { 
     # faster evaluation in this case
     exp_x_cumprod <- aperm(
-      array(apply(exp(x), marg_othdim, cumprod), dim = c(dim_thres, dim_noncat)),
+      array(apply(exp(x), marg_othdim, cumprod), dim = dim_t),
       perm = c(marg_othdim + 1, 1)
     )
     out <- abind::abind(ones_arr, exp_x_cumprod)
   } else {
     x <- ilink(x, link)
     x_cumprod <- aperm(
-      array(apply(x, marg_othdim, cumprod), dim = c(dim_thres, dim_noncat)),
+      array(apply(x, marg_othdim, cumprod), dim = dim_t),
       perm = c(marg_othdim + 1, 1)
     )
     nthres <- dim(x)[ndim]
+    Sx_cumprod_rev <- apply(
+      1 - slice(x, ndim, rev(seq_len(nthres)), drop = FALSE),
+      marg_othdim, cumprod
+    )
     Sx_cumprod_rev <- aperm(
-      array(apply(
-        1 - slice(x, ndim, rev(seq_len(nthres)), drop = FALSE),
-        marg_othdim, cumprod
-      ), dim = c(dim_thres, dim_noncat)), 
+      array(Sx_cumprod_rev, dim = dim_t), 
       perm = c(marg_othdim + 1, 1)
     )
     Sx_cumprod_rev <- slice(