--- a/tests/testthat/helpers/d_ordinal_sim.R
+++ b/tests/testthat/helpers/d_ordinal_sim.R
@@ -1,4 +1,4 @@
 # This test corresponds to a single observation.
 set.seed(1234)
-ndraws <- 5
+ndraws_vec <- c(1, 5)
 ncat_vec <- c(2, 3)