--- a/tests/testthat/helpers/inv_link_ordinal_sim.R
+++ b/tests/testthat/helpers/inv_link_ordinal_sim.R
@@ -1,9 +1,4 @@
 set.seed(1234)
 ndraws <- 5
 nobsv <- 4
-ncat <- 3
-x_test <- array(rnorm(ndraws * nobsv * (ncat - 1)),
-                dim = c(ndraws, nobsv, ncat - 1))
-nx_test <- -x_test
-exp_nx_cumprod <- aperm(apply(exp(nx_test), c(1, 2), cumprod),
-                        perm = c(2, 3, 1))
+ncat_vec <- c(2, 3)