--- a/tests/testthat/helpers/inv_link_ordinal_sim.R
+++ b/tests/testthat/helpers/inv_link_ordinal_sim.R
@@ -1,4 +1,4 @@
 set.seed(1234)
 ndraws_vec <- c(1, 5)
-nobsv <- 4
+nobsv_vec <- c(1, 4)
 ncat_vec <- c(2, 3)