This PR is based on PR #1154, so it's better to merge #1154 first.

As suspected [here](https://github.com/paul-buerkner/brms/pull/1137#issuecomment-832596346) (enumeration point 2), the current unit tests for the `inv_link_<ordinal_family>()` functions do indeed provide a more efficient implementation of these `inv_link_<ordinal_family>()` functions. Here is my requested speed comparison:
```r
library(brms)
options(mc.cores = parallel::detectCores(logical = FALSE))
data(inhaler, package = "brms")
bfit_cumul <- brm(
  formula = rating ~ period + carry + treat + (1 | subject),
  data = inhaler,
  family = cumulative(),
  seed = 475064792
)
library(microbenchmark)
microbenchmark(epred_cumul <- posterior_epred(bfit_cumul),
               times = 25)
### Old:
# Unit: seconds
#                                       expr      min       lq     mean   median      uq      max neval
# epred_cumul <- posterior_epred(bfit_cumul) 1.206015 1.415697 1.525062 1.429557 1.68592 2.037138    25
### 
### New:
# Unit: seconds
#                                       expr      min       lq    mean   median       uq      max neval
# epred_cumul <- posterior_epred(bfit_cumul) 1.322469 1.329213 1.35928 1.336847 1.343451 1.615867    25
### 

bfit_sratio <- update(bfit_cumul, family = sratio())
microbenchmark(epred_sratio <- posterior_epred(bfit_sratio),
               times = 25)
### Old:
# Unit: seconds
#                                         expr      min       lq     mean   median       uq      max neval
# epred_sratio <- posterior_epred(bfit_sratio) 13.98803 14.29517 14.40808 14.38126 14.58076 14.65054    25
### 
### New:
# Unit: seconds
#                                         expr      min       lq     mean   median       uq      max neval
# epred_sratio <- posterior_epred(bfit_sratio) 8.475701 8.635784 8.723748 8.690884 8.795831 9.093068    25
### 

bfit_cratio <- update(bfit_cumul, family = cratio())
microbenchmark(epred_cratio <- posterior_epred(bfit_cratio),
               times = 25)
### Old:
# Unit: seconds
#                                         expr      min       lq     mean   median       uq      max neval
# epred_cratio <- posterior_epred(bfit_cratio) 13.97217 14.32164 14.39963 14.34807 14.47658 14.70686    25
### 
### New:
# Unit: seconds
#                                         expr      min       lq    mean   median       uq      max neval
# epred_cratio <- posterior_epred(bfit_cratio) 8.169052 8.572888 8.62483 8.597578 8.634792 8.878164    25
### 

bfit_acat <- update(bfit_cumul, family = acat())
microbenchmark(epred_acat <- posterior_epred(bfit_acat),
               times = 25)
### Old:
# Unit: seconds
#                                     expr      min       lq     mean   median       uq      max neval
# epred_acat <- posterior_epred(bfit_acat) 13.73694 13.88184 13.94854 13.92991 13.95867 14.17641    25
### 
### New:
# Unit: seconds
#                                     expr      min       lq    mean   median       uq      max neval
# epred_acat <- posterior_epred(bfit_acat) 12.32044 12.65225 12.7692 12.72448 12.90858 13.42917    25
### 

bfit_acat_probit <- update(bfit_cumul, family = acat(link = "probit"))
microbenchmark(epred_acat_probit <- posterior_epred(bfit_acat_probit),
               times = 25)
### Old:
# Unit: seconds
#                                                   expr      min      lq     mean   median       uq      max neval
# epred_acat_probit <- posterior_epred(bfit_acat_probit) 31.17294 31.3034 31.45372 31.49983 31.55278 31.70364    25
### 
### New:
# Unit: seconds
#                                                   expr      min       lq     mean   median       uq      max neval
# epred_acat_probit <- posterior_epred(bfit_acat_probit) 21.54279 22.21362 22.39333 22.48148 22.65841 23.15681    25
### 

```

Because of this speed improvement (which is sometimes smaller, sometimes larger, but always present), this PR swaps the two implementations of the `inv_link_<ordinal_family>()` functions (the original one and the one from the unit tests).

Of course, one could go one step further and achieve another speed improvement by using arrays when calling `d<ordinal_family>()` in `posterior_epred_ordinal()`, i.e. by not iterating over the observations, but instead including them as an additional array dimension. But that probably requires larger changes.