@paul-buerkner 2021-05-06
Very elegant implementations. Thank you! Will merge once the checks pass.