@paul-buerkner 2021-05-06
When I change the number of categories in `inv_link_ordinal_sim` from `3` to `2` some tests fail. Can you take a look at fix the implementations to work with 2 categories as well?