@fweber144 2021-05-06
Should be fixed now. And I hope I have included all special cases in the unit tests now. Thanks again for pointing this out and sorry for not being aware of this.