@paul-buerkner 2021-05-06
Thanks for fixing this! And no worries, 1/3 of all brms bugs are caused by R dropping dimensions somewhere in edge cases :-D