@fweber144 2021-05-07
Yeah, that dropping of dimensions is not really developer-friendly, especially when there's no way to turn it off, as for `apply()`. Thanks for merging!