@jgabry 2021-05-07
> And no worries, 1/3 of all brms bugs are caused by R dropping dimensions somewhere in edge cases :-D

1/3 of all nightmares I have are caused by R dropping dimensions ;)