@fweber144 2021-05-08
You mean because I repeatedly added those `array(..., dim = c(dim_thres, dim_noncat))` calls in commit 8b1704ac0897cae987e3f1d2909e7a20c62051f5? That's true, that could perhaps have been solved more elegantly by adding a custom wrapper around `apply()` which does not drop margins.