@paul-buerkner 2021-05-08
that was no comment about your code I think but only about Rs way to
filling arrays with more numbers from the start if the sizes don't match.

 although you are right, we could add a apply2 function as a wrapper around
apply.

Frank Weber ***@***.***> schrieb am Sa., 8. Mai 2021, 09:23:

> You mean because I repeatedly added those array(..., dim = c(dim_thres,
> dim_noncat)) calls in commit 8b1704a
> <https://github.com/paul-buerkner/brms/commit/8b1704ac0897cae987e3f1d2909e7a20c62051f5>?
> That's true, that could perhaps have been solved more elegantly by adding a
> custom wrapper around apply() which does not drop margins.
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/paul-buerkner/brms/pull/1155#issuecomment-835168654>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ADCW2AEAU6E4WPMDR2SJ6WDTMTRFVANCNFSM44G6RHYQ>
> .
>