@fweber144 2021-05-08
Ah I see :D

Concerning the `apply()` wrapper: I'm currently lacking the time to implement this, but I'll try to keep it in mind.