@paul-buerkner 2021-05-08
yeah no worries. I can just do that myself. the projpred things are much
more important anyway (and probably a lot of other things you have on your
table).

Frank Weber ***@***.***> schrieb am Sa., 8. Mai 2021, 17:54:

> Ah I see :D
>
> Concerning the apply() wrapper: I'm currently lacking the time to
> implement this, but I'll try to keep it in mind.
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/paul-buerkner/brms/pull/1155#issuecomment-835410434>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ADCW2AB4NPWJV76U6KTLOH3TMVNENANCNFSM44G6RHYQ>
> .
>