@fweber144 2021-05-20
As if the R Core Team had heard us: For R 4.1.0, the [NEWS file](https://cran.r-project.org/doc/manuals/r-release/NEWS.html) says:

> apply() gains a simplify argument to allow disabling of simplification of results.

This new argument doesn't offer exactly what I would have desired, but it should be a good starting point for writing a custom `apply()` wrapper. @paul-buerkner, do you want me to write such a wrapper? But it would make brms depend on R >= 4.1.0.