@paul-buerkner 2021-05-20
thanks for looking into it. I don't want to strictly depend on R 4.1+ or
even R 4.0+ at the moment so I suggest we don't change it for now.

Frank Weber ***@***.***> schrieb am Do., 20. Mai 2021, 06:09:

> As if the R Core Team had heard us: For R 4.1.0, the NEWS file
> <https://cran.r-project.org/doc/manuals/r-release/NEWS.html> says:
>
> apply() gains a simplify argument to allow disabling of simplification of
> results.
>
> This new argument doesn't offer exactly what I would have desired, but it
> should be a good starting point for writing a custom apply() wrapper.
> @paul-buerkner <https://github.com/paul-buerkner>, do you want me to
> write such a wrapper? But it would make brms depend on R >= 4.1.0.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/paul-buerkner/brms/pull/1155#issuecomment-844671110>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ADCW2AGHTDBOEFV53N4WYKTTOSDRLANCNFSM44G6RHYQ>
> .
>